Cleanup to prepare for review.
This moves argument parsing into its own file, moves some methods
into CocoaPodUtils and consolidates some related types.
