+++ ZipBuilder/Sources/ZipBuilder/CocoaPodUtils.swift
@@ -13,73 +13,32 @@ public enum CocoaPodUtils {
 
   /// Information associated with an installed pod.
   public struct PodInfo {
+    /// Public name of the pod.
     var name: String
+
+    /// The version of the pod.
     var version: String
-    var location: URL
-  }
 
-  /// Generates a key representing the unique combination of all subspecs used for that Pod. This is
-  /// necessary for Pods like GoogleUtilities, where we will need to include all subspecs as part of
-  /// a build. Otherwise we could accidentally use a cached framework that doesn't include all the
-  /// code necessary to function.
-  ///
-  /// - Parameters:
-  ///   - framework: The framework being built.
-  ///   - podfileLock: The contents of the Podfile.lock for the project.
-  /// - Returns: A key to describe the full set of subspecs used to build the framework, or an empty
-  ///            String if there were no specific subspecs used.
-  public static func cacheKey(forPod podName: String,
-                              fromPodfileLock podfileLock: String) -> String? {
-    // Get the first section of the Podfile containing only Pods installed, the only thing we care
-    // about.
-    guard let podsInstalled = podfileLock.components(separatedBy: "DEPENDENCIES:").first else {
-      fatalError("""
-        Could not generate cache key for \(podName) from Podfile.lock contents - is this a valid
-        Podfile.lock?
-        ---------- Podfile.lock contents ----------
-        \(podfileLock)
-        -------------------------------------------
-        """)
+    /// The location of the pod on disk.
+    var installedLocation: URL
+
+    /// A key that can be generated and used for identifying pods due to binary differences.
+    var cacheKey: String?
+
+    /// Default initializer. Explicitly declared to take advantage of default arguments.
+    init(name: String, version: String, installedLocation: URL, cacheKey: String? = nil) {
+      self.name = name
+      self.version = version
+      self.installedLocation = installedLocation
+      self.cacheKey = cacheKey
     }
-
-    // Only get the lines that start with "- ", and have the framework we're looking for since they
-    // are the top level pods that are installed.
-    // Example result of a single line: `- GoogleUtilities/Environment (~> 5.2)`.
-    let lines = podsInstalled.components(separatedBy: .newlines).filter {
-      return $0.hasPrefix("- ") && $0.contains(podName)
-    }
-
-    // Get a list of all the subspecs used to build this framework, and use that to generate the
-    // cache key.
-    var uniqueSubspecs = Set<String>()
-    for line in lines.sorted() {
-      // Separate the line into readable chunks, using a space and quote as a separator.
-      // Example result: `["-", "GoogleUtilities/Environment", "(~>", "5.2)"]`.
-      let components = line.components(separatedBy: CharacterSet(charactersIn: " \""))
-
-      // The Pod and subspec will be the only variables we care about, filter out the rest of the
-      // components.
-      // Example result: 'GoogleUtilities/Environment'.
-      guard let fullPodName = components.filter({ $0.contains(podName) }).first else { continue }
-
-      // The fullPodName will be something like `GoogleUtilities/UserDefaults`.
-      let subspec = fullPodName.replacingOccurrences(of: "\(podName)/", with: "")
-      if !subspec.isEmpty {
-        uniqueSubspecs.insert(subspec)
-      }
-    }
-
-    // Return nil if there are no subpsecs used, since no cache key is necessary.
-    guard !uniqueSubspecs.isEmpty else {
-      return nil
-    }
-
-    // Assemble the cache key based on the framework name, and all subspecs (sorted alphabetically
-    // for repeatability) separated by a `+` (as was previously used).
-    return podName + "+" + uniqueSubspecs.sorted().joined(separator: "+")
   }
 
   /// Execute the `pod cache list` command to get the Pods currently cached on your machine.
+  /// Executes the `pod cache list` command to get the Pods curerntly cached on your machine.
+  ///
+  /// - Parameter dir: The directory containing all installed pods.
+  /// - Returns: A dictionary keyed by the pod name, then by version number.
   public static func listPodCache(inDir dir: URL) -> [String: [String: PodInfo]] {
     let result = Shell.executeCommandFromScript("pod cache list", outputToConsole: false)
     switch result {
@@ -91,8 +50,48 @@ public enum CocoaPodUtils {
     }
   }
 
-  /// Install an array of subspecs from the Firebase pod in a specific directory.
-  public static func installSubspecs(_ subspecs: [Subspec], inDir directory: URL) {
+  /// Gets metadata from installed Pods. Reads the `Podfile.lock` file and parses it.
+  public static func installedPodsInfo(inProjectDir projectDir: URL) -> [PodInfo] {
+    // Read from the Podfile.lock to get the installed versions and names.
+    let podfileLock: String
+    do {
+      podfileLock = try String(contentsOf: projectDir.appendingPathComponent("Podfile.lock"))
+    } catch {
+      fatalError("Could not read contents of `Podfile.lock` to get installed Pod info in " +
+        "\(projectDir): \(error)")
+    }
+
+    // Get the versions in the format of [PodName: VersionString].
+    let versions = self.loadVersionsFromPodfileLock(contents: podfileLock)
+
+    // Generate an InstalledPod for each Pod found.
+    let podsDir = projectDir.appendingPathComponent("Pods")
+    var installedPods: [PodInfo] = []
+    for (podName, version) in versions {
+      let podDir = podsDir.appendingPathComponent(podName)
+      guard FileManager.default.directoryExists(at: podDir) else {
+        fatalError("Directory for \(podName) doesn't exist at \(podDir) - failed while getting " +
+          "information for installed Pods.")
+      }
+
+      // Generate the cache key for this framework. We will use the list of subspecs used in the Pod
+      // to generate this, since a Pod like GoogleUtilities could build different sources based on
+      // what subspecs are included.
+      let cacheKey = self.cacheKey(forPod: podName, fromPodfileLock: podfileLock)
+      let podInfo = PodInfo(name: podName,
+                            version: version,
+                            installedLocation: podDir,
+                            cacheKey: cacheKey)
+      installedPods.append(podInfo)
+    }
+
+    return installedPods
+  }
+
+  /// Install an array of subspecs from the Firebase pod in a specific directory, returning an array
+  /// of PodInfo for each pod that was installed.
+  @discardableResult
+  public static func installSubspecs(_ subspecs: [Subspec], inDir directory: URL) -> [PodInfo] {
     let fileManager = FileManager.default
     // Ensure the directory exists, otherwise we can't install all subspecs.
     guard fileManager.directoryExists(at: directory) else {
@@ -127,8 +126,9 @@ public enum CocoaPodUtils {
                 \(output)
                 """)
     case .success(let output):
-      // Print the output to the console.
+      // Print the output to the console and return the information for all installed pods.
       print(output)
+      return self.installedPodsInfo(inProjectDir: directory)
     }
   }
 
@@ -202,6 +202,67 @@ public enum CocoaPodUtils {
     return (framework, version)
   }
 
+  /// Generates a key representing the unique combination of all subspecs used for that Pod. This is
+  /// necessary for Pods like GoogleUtilities, where we will need to include all subspecs as part of
+  /// a build. Otherwise we could accidentally use a cached framework that doesn't include all the
+  /// code necessary to function.
+  ///
+  /// - Parameters:
+  ///   - framework: The framework being built.
+  ///   - podfileLock: The contents of the Podfile.lock for the project.
+  /// - Returns: A key to describe the full set of subspecs used to build the framework, or an empty
+  ///            String if there were no specific subspecs used.
+  private static func cacheKey(forPod podName: String,
+                               fromPodfileLock podfileLock: String) -> String? {
+    // Get the first section of the Podfile containing only Pods installed, the only thing we care
+    // about.
+    guard let podsInstalled = podfileLock.components(separatedBy: "DEPENDENCIES:").first else {
+      fatalError("""
+        Could not generate cache key for \(podName) from Podfile.lock contents - is this a valid
+        Podfile.lock?
+        ---------- Podfile.lock contents ----------
+        \(podfileLock)
+        -------------------------------------------
+        """)
+    }
+
+    // Only get the lines that start with "- ", and have the framework we're looking for since they
+    // are the top level pods that are installed.
+    // Example result of a single line: `- GoogleUtilities/Environment (~> 5.2)`.
+    let lines = podsInstalled.components(separatedBy: .newlines).filter {
+      return $0.hasPrefix("- ") && $0.contains(podName)
+    }
+
+    // Get a list of all the subspecs used to build this framework, and use that to generate the
+    // cache key.
+    var uniqueSubspecs = Set<String>()
+    for line in lines.sorted() {
+      // Separate the line into readable chunks, using a space and quote as a separator.
+      // Example result: `["-", "GoogleUtilities/Environment", "(~>", "5.2)"]`.
+      let components = line.components(separatedBy: CharacterSet(charactersIn: " \""))
+
+      // The Pod and subspec will be the only variables we care about, filter out the rest of the
+      // components.
+      // Example result: 'GoogleUtilities/Environment'.
+      guard let fullPodName = components.filter({ $0.contains(podName) }).first else { continue }
+
+      // The fullPodName will be something like `GoogleUtilities/UserDefaults`.
+      let subspec = fullPodName.replacingOccurrences(of: "\(podName)/", with: "")
+      if !subspec.isEmpty {
+        uniqueSubspecs.insert(subspec)
+      }
+    }
+
+    // Return nil if there are no subpsecs used, since no cache key is necessary.
+    guard !uniqueSubspecs.isEmpty else {
+      return nil
+    }
+
+    // Assemble the cache key based on the framework name, and all subspecs (sorted alphabetically
+    // for repeatability) separated by a `+` (as was previously used).
+    return podName + "+" + uniqueSubspecs.sorted().joined(separator: "+")
+  }
+
   /// Create the contents of a Podfile for an array of subspecs. This assumes the array of subspecs
   /// is not empty.
   private static func generatePodfile(for subspecs: [Subspec]) -> String {
@@ -258,7 +319,7 @@ public enum CocoaPodUtils {
           break;
         case "Pod":
           let podLocation = URL(fileURLWithPath: value)
-          let podInfo = PodInfo(name: podName!, version: podVersion!, location: podLocation)
+          let podInfo = PodInfo(name: podName!, version: podVersion!, installedLocation: podLocation)
           if podsCache[podName!] == nil {
             podsCache[podName!] = [:]
           }
