+++ ZipBuilder/Sources/ZipBuilder/main.swift
@@ -1,185 +1,5 @@
 import Foundation
-
-/// Describes an object that can check if a file exists in the filesystem. Used to allow for better
-/// testing with FileManager.
-protocol FileChecker {
-  /// Returns a Boolean value that indicates whether a file or directory exists at a specified path.
-  func fileExists(atPath: String) -> Bool
-}
-
-// Make FileManager a FileChecker. This is empty since FileManager already provides this function.
-extension FileManager: FileChecker {}
-
-/// LaunchArgs reads from UserDefaults to assemble all launch arguments coming from the command line
-/// or the Xcode scheme. UserDefaults contains all launch arguments that are in the format of:
-/// `-myKey myValue`.
-struct LaunchArgs {
-  /// Keys associated with the launch args. See `Usage` for descriptions of each flag.
-  private enum Key: String {
-    case cacheEnabled
-    case customSpecRepos
-    case debugMode
-    case deleteCache
-    case existingVersions
-    case releasingSDKs
-    case templateDir
-    case updatePodRepo
-
-    /// Usage description for the key.
-    var usage: String {
-      switch self {
-      case .cacheEnabled:
-        return "A flag to control using the cache for frameworks."
-      case .customSpecRepos:
-        return "A comma separated list of custom CocoaPod Spec repos."
-      case .debugMode:
-        return "A flag to enable debug mode. Pass in `true/false`, `YES/NO`, or `1/0`."
-      case .deleteCache:
-        return "A flag to empty the cache. Note: if this flag and the `cacheEnabled` flag is " +
-        "set, it will fail since that's probably unintended."
-      case .existingVersions:
-        return "The file path to a textproto file containing the existing released SDK versions, " +
-               "of type `ZipBuilder_FirebaseSDKs`."
-      case .releasingSDKs:
-        return " The file path to a textproto file containing all the releasing SDKs, of type " +
-               "`ZipBuilder_Release`."
-      case .templateDir:
-        return "The path to the directory containing the blank xcodeproj and Info.plist for " +
-               "building source based frameworks"
-      case .updatePodRepo:
-        return "A flag to run `pod repo update` before building the zip file."
-      }
-    }
-
-    // TODO: Once we default to Swift 4.2 (in Xcode 10) we can conform to "CaseIterable" protocol to
-    //       automatically generate this method.
-    /// All the subspecs to parse.
-    static func allCases() -> [Key] {
-      return [.cacheEnabled,
-              .customSpecRepos,
-              .debugMode,
-              .deleteCache,
-              .existingVersions,
-              .releasingSDKs,
-              .templateDir,
-              .updatePodRepo]
-    }
-  }
-
-  /// A file URL to a textproto with the contents of a `ZipBuilder_FirebaseSDKs` object. Used to
-  /// verify expected version numbers.
-  let allSDKsPath: URL?
-
-  /// A file URL to a textproto with the contents of a `ZipBuilder_Release` object. Used to verify
-  /// expected version numbers.
-  let currentReleasePath: URL?
-
-  /// Custom CocoaPods spec repos to be used. If not provided, the tool will only use the CocoaPods
-  /// master repo.
-  let customSpecRepos: [URL]?
-
-  /// The path to the directory containing the blank xcodeproj and Info.plist for building source
-  /// based frameworks.
-  let templateDir: URL
-
-  /// A flag to control using the cache for frameworks.
-  let cacheEnabled: Bool
-
-  /// A flag to delete the cache from the cache directory.
-  let deleteCache: Bool
-
-  /// A flag to update the Pod Repo or not.
-  let updatePodRepo: Bool
-
-  /// Initializes with values pulled from the instance of UserDefaults passed in.
-  ///
-  /// - Parameters:
-  ///   - defaults: User defaults containing launch arguments. Defaults to `standard`.
-  ///   - fileChecker: An object that can check if a file exists or not. Defaults to
-  ///                  `FileManager.default`.
-  init(userDefaults defaults: UserDefaults = UserDefaults.standard,
-       fileChecker: FileChecker = FileManager.default) {
-    // Get the project template directory, and fail if it doesn't exist.
-    guard let templatePath = defaults.string(forKey: Key.templateDir.rawValue) else {
-      LaunchArgs.exitWithUsageAndLog("Missing \(Key.templateDir) key for the folder containing " +
-        "all required files to build frameworks.")
-    }
-
-    templateDir = URL(fileURLWithPath: templatePath)
-
-    // Parse the existing versions key.
-    if let existingVersions = defaults.string(forKey: Key.existingVersions.rawValue) {
-      let url = URL(fileURLWithPath: existingVersions)
-      guard fileChecker.fileExists(atPath: url.path) else {
-        LaunchArgs.exitWithUsageAndLog("Could not parse \(Key.existingVersions) key: value " +
-          "passed in is not a file URL or the file does not exist. Value: \(existingVersions)")
-      }
-
-      allSDKsPath = url.standardizedFileURL
-    } else {
-      // No argument was passed in.
-      allSDKsPath = nil
-    }
-
-    // Parse the current releases key.
-    if let currentRelease = defaults.string(forKey: Key.releasingSDKs.rawValue) {
-      let url = URL(fileURLWithPath: currentRelease)
-      guard fileChecker.fileExists(atPath: url.path) else {
-        LaunchArgs.exitWithUsageAndLog("Could not parse \(Key.releasingSDKs) key: value passed " +
-          "in is not a file URL or the file does not exist. Value: \(currentRelease)")
-      }
-
-      currentReleasePath = url.standardizedFileURL
-    } else {
-      // No argument was passed in.
-      currentReleasePath =  nil
-    }
-
-    // Parse the custom specs key.
-    if let customSpecs = defaults.string(forKey: Key.customSpecRepos.rawValue) {
-      // Custom specs are passed in as a comma separated list of URLs. Split the String by each
-      // comma and map it to URLs. If any URL is invalid, fail immediately.
-      let specs = customSpecs.split(separator: ",").map { (specStr: Substring) -> URL in
-        guard let spec = URL(string: String(specStr)) else {
-          LaunchArgs.exitWithUsageAndLog("Error parsing specs: \(specStr) is not a valid URL.")
-        }
-
-        return spec
-      }
-
-     customSpecRepos = specs
-    } else {
-      // No argument was passed in.
-      customSpecRepos = nil
-    }
-
-    updatePodRepo = defaults.bool(forKey: Key.updatePodRepo.rawValue)
-
-    // Parse the cache keys. If no value is provided for each, it defaults to `false`.
-    cacheEnabled = defaults.bool(forKey: Key.cacheEnabled.rawValue)
-    deleteCache = defaults.bool(forKey: Key.deleteCache.rawValue)
-
-    if deleteCache && cacheEnabled {
-      LaunchArgs.exitWithUsageAndLog("Invalid pair - attempted to delete the cache and enable " +
-                                     "it at the same time. Please remove on of the keys and try " +
-                                     "again.")
-    }
-  }
-
-  /// Prints an error that occurred, the proper usage String, and quits the application.
-  private static func exitWithUsageAndLog(_ errorText: String) -> Never {
-    print(errorText)
-    print("Usage: `swift run ZipBuilder [ARGS]` where args are:")
-    for option in Key.allCases() {
-      print("""
-        -\(option.rawValue) <VALUE>
-                \(option.usage)
-        """)
-    }
-    // Loop over all the possible keys and print their description.
-    fatalError("Invalid arguments. See above for specific error and usage instructions.")
-  }
-}
+import Utility
 
 // Get the launch arguments, parsed by user defaults.
 let args = LaunchArgs()
@@ -194,9 +14,14 @@ if args.deleteCache {
   }
 }
 
+// Keep timing for how long it takes to build the zip file for information purposes.
+let buildStart = Date()
+var cocoaPodsUpdateMessage: String = ""
+
 // Do a Pod Update if requested.
 if args.updatePodRepo {
   CocoaPodUtils.updateRepos()
+  cocoaPodsUpdateMessage = "CocoaPods took \(-buildStart.timeIntervalSinceNow) seconds to update."
 }
 
 let builder = ZipBuilder(templateDir: args.templateDir,
@@ -206,8 +31,24 @@ let builder = ZipBuilder(templateDir: args.templateDir,
                          useCache: args.cacheEnabled)
 
 do {
-  let location = try builder.build()
+  // Build the zip file and get the path.
+  let location = try builder.buildAndAssembleZipDir()
+
+  // Get the time since the start of the build to get the full time.
+  let secondsSinceStart = -Int(buildStart.timeIntervalSinceNow)
+  print("""
+    Time profile:
+      It took \(secondsSinceStart) seconds (~\(secondsSinceStart / 60)m) to build the zip file.
+      \(cocoaPodsUpdateMessage)
+    """)
+
   print("Location of zip file: \(location)")
 } catch let error {
+  let secondsSinceStart = -buildStart.timeIntervalSinceNow
+  print("""
+    Time profile:
+      The build failed in \(secondsSinceStart) seconds (~\(secondsSinceStart / 60)m).
+      \(cocoaPodsUpdateMessage)
+    """)
   fatalError("Could not build the zip file: \(error)")
 }
